rename simulate to simulate_msa
- Rename substitution simulator module file to
- Update module export in  to use
- Preserve tests and functionality while switching filename to better reflect purpose
